+++ partiql-lang/src/main/kotlin/org/partiql/lang/eval/EvaluatingCompiler.kt
@@ -92,7 +92,7 @@ import java.util.regex.Pattern
  *
  * This name was chosen because it is a thunk that accepts an instance of `Environment`.
  */
-private typealias ThunkEnv = Thunk<Environment>
+internal typealias ThunkEnv = Thunk<Environment>
 
 /**
  * A thunk taking a single [T] argument and the current environment.
@@ -125,18 +125,18 @@ private typealias ThunkEnvValue<T> = ThunkValue<Environment, T>
  * @param functions A map of functions keyed by function name that will be available during compilation.
  * @param compileOptions Various options that effect how the source code is compiled.
  */
-internal class EvaluatingCompiler(
+internal open class EvaluatingCompiler(
     private val functions: List<ExprFunction>,
     private val customTypedOpParameters: Map<String, TypedOpParameter>,
     private val procedures: Map<String, StoredProcedure>,
-    private val compileOptions: CompileOptions = CompileOptions.standard()
+    internal val compileOptions: CompileOptions = CompileOptions.standard()
 ) {
 
     // TODO: remove this once we migrate from `IonValue` to `IonElement`.
     private val ion = IonSystemBuilder.standard().build()
 
     private val errorSignaler = compileOptions.typingMode.createErrorSignaler()
-    private val thunkFactory = compileOptions.typingMode.createThunkFactory<Environment>(compileOptions.thunkOptions)
+    internal val thunkFactory = compileOptions.typingMode.createThunkFactory<Environment>(compileOptions.thunkOptions)
     private val functionManager = FunctionManager(functions)
 
     private val compilationContextStack = Stack<CompilationContext>()
@@ -152,7 +152,7 @@ internal class EvaluatingCompiler(
 
     // Note: please don't make this inline -- it messes up [EvaluationException] stack traces and
     // isn't a huge benefit because this is only used at SQL-compile time anyway.
-    private fun <R> nestCompilationContext(
+    internal fun <R> nestCompilationContext(
         expressionContext: ExpressionContext,
         fromSourceNames: Set<String>,
         block: () -> R
@@ -348,7 +348,7 @@ internal class EvaluatingCompiler(
      * and throws [InterruptedException] if [Thread.interrupted] it has been set in the
      * hope that long-running compilations may be aborted by the caller.
      */
-    fun compile(originalAst: PartiqlAst.Statement): Expression {
+    open fun compile(originalAst: PartiqlAst.Statement): Expression {
         val visitorTransform = compileOptions.visitorTransformMode.createVisitorTransform()
         val transformedAst = visitorTransform.transformStatement(originalAst)
         val partiqlAstSanityValidator = PartiqlAstSanityValidator()
@@ -360,15 +360,25 @@ internal class EvaluatingCompiler(
         }
 
         return object : Expression {
-            override fun eval(session: EvaluationSession): ExprValue {
+            override val coverageStructure: CoverageStructure? = null
 
+            override fun eval(session: EvaluationSession): ExprValue {
                 val env = Environment(
                     session = session,
                     locals = session.globals,
                     current = session.globals
                 )
-
                 return thunk(env)
+            }
+
+            override fun evaluate(session: EvaluationSession): PartiQLResult {
+                val env = Environment(
+                    session = session,
+                    locals = session.globals,
+                    current = session.globals
+                )
+                val value = thunk(env)
+                return PartiQLResult.Value(value = value)
             }
         }
     }
@@ -383,7 +393,7 @@ internal class EvaluatingCompiler(
      *
      * This function will [InterruptedException] if [Thread.interrupted] has been set.
      */
-    private fun compileAstStatement(ast: PartiqlAst.Statement): ThunkEnv {
+    internal open fun compileAstStatement(ast: PartiqlAst.Statement): ThunkEnv {
         checkThreadInterrupted()
         return when (ast) {
             is PartiqlAst.Statement.Query -> compileAstExpr(ast.expr)
@@ -398,7 +408,7 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileAstExpr(expr: PartiqlAst.Expr): ThunkEnv {
+    internal open fun compileAstExpr(expr: PartiqlAst.Expr): ThunkEnv {
         val metas = expr.metas
 
         return when (expr) {
@@ -713,7 +723,7 @@ internal class EvaluatingCompiler(
         return checkIntegerOverflow(computeThunk, metas)
     }
 
-    private fun compileEq(expr: PartiqlAst.Expr.Eq, metas: MetaContainer): ThunkEnv {
+    internal open fun compileEq(expr: PartiqlAst.Expr.Eq, metas: MetaContainer): ThunkEnv {
         val argThunks = compileAstExprs(expr.operands)
 
         return thunkFactory.thunkAndMap(metas, argThunks) { lValue, rValue ->
@@ -721,7 +731,7 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileNe(expr: PartiqlAst.Expr.Ne, metas: MetaContainer): ThunkEnv {
+    internal open fun compileNe(expr: PartiqlAst.Expr.Ne, metas: MetaContainer): ThunkEnv {
         val argThunks = compileAstExprs(expr.operands)
 
         return thunkFactory.thunkFold(metas, argThunks) { lValue, rValue ->
@@ -729,31 +739,31 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileLt(expr: PartiqlAst.Expr.Lt, metas: MetaContainer): ThunkEnv {
+    internal open fun compileLt(expr: PartiqlAst.Expr.Lt, metas: MetaContainer): ThunkEnv {
         val argThunks = compileAstExprs(expr.operands)
 
         return thunkFactory.thunkAndMap(metas, argThunks) { lValue, rValue -> lValue < rValue }
     }
 
-    private fun compileLte(expr: PartiqlAst.Expr.Lte, metas: MetaContainer): ThunkEnv {
+    internal open fun compileLte(expr: PartiqlAst.Expr.Lte, metas: MetaContainer): ThunkEnv {
         val argThunks = compileAstExprs(expr.operands)
 
         return thunkFactory.thunkAndMap(metas, argThunks) { lValue, rValue -> lValue <= rValue }
     }
 
-    private fun compileGt(expr: PartiqlAst.Expr.Gt, metas: MetaContainer): ThunkEnv {
+    internal open fun compileGt(expr: PartiqlAst.Expr.Gt, metas: MetaContainer): ThunkEnv {
         val argThunks = compileAstExprs(expr.operands)
 
         return thunkFactory.thunkAndMap(metas, argThunks) { lValue, rValue -> lValue > rValue }
     }
 
-    private fun compileGte(expr: PartiqlAst.Expr.Gte, metas: MetaContainer): ThunkEnv {
+    internal open fun compileGte(expr: PartiqlAst.Expr.Gte, metas: MetaContainer): ThunkEnv {
         val argThunks = compileAstExprs(expr.operands)
 
         return thunkFactory.thunkAndMap(metas, argThunks) { lValue, rValue -> lValue >= rValue }
     }
 
-    private fun compileBetween(expr: PartiqlAst.Expr.Between, metas: MetaContainer): ThunkEnv {
+    internal open fun compileBetween(expr: PartiqlAst.Expr.Between, metas: MetaContainer): ThunkEnv {
         val valueThunk = compileAstExpr(expr.value)
         val fromThunk = compileAstExpr(expr.from)
         val toThunk = compileAstExpr(expr.to)
@@ -779,7 +789,7 @@ internal class EvaluatingCompiler(
      * `IN` is varies from the `OR` operator in that this behavior holds true when other types of expressions are
      * used on the right side of `IN` such as sub-queries and variables whose value is that of a list or bag.
      */
-    private fun compileIn(expr: PartiqlAst.Expr.InCollection, metas: MetaContainer): ThunkEnv {
+    internal open fun compileIn(expr: PartiqlAst.Expr.InCollection, metas: MetaContainer): ThunkEnv {
         val args = expr.operands
         val leftThunk = compileAstExpr(args[0])
         val rightOp = args[1]
@@ -866,7 +876,7 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileNot(expr: PartiqlAst.Expr.Not, metas: MetaContainer): ThunkEnv {
+    internal open fun compileNot(expr: PartiqlAst.Expr.Not, metas: MetaContainer): ThunkEnv {
         val argThunk = compileAstExpr(expr.expr)
 
         return thunkFactory.thunkEnvOperands(metas, argThunk) { _, value ->
@@ -874,7 +884,7 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileAnd(expr: PartiqlAst.Expr.And, metas: MetaContainer): ThunkEnv {
+    internal open fun compileAnd(expr: PartiqlAst.Expr.And, metas: MetaContainer): ThunkEnv {
         val argThunks = compileAstExprs(expr.operands)
 
         // can't use the null propagation supplied by [ThunkFactory.thunkEnv] here because AND short-circuits on
@@ -919,7 +929,7 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileOr(expr: PartiqlAst.Expr.Or, metas: MetaContainer): ThunkEnv {
+    internal open fun compileOr(expr: PartiqlAst.Expr.Or, metas: MetaContainer): ThunkEnv {
         val argThunks = compileAstExprs(expr.operands)
 
         // can't use the null propagation supplied by [ThunkFactory.thunkEnv] here because OR short-circuits on
@@ -992,7 +1002,7 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileCall(expr: PartiqlAst.Expr.Call, metas: MetaContainer): ThunkEnv {
+    internal open fun compileCall(expr: PartiqlAst.Expr.Call, metas: MetaContainer): ThunkEnv {
         val funcArgThunks = compileAstExprs(expr.args)
         val arity = funcArgThunks.size
         val name = expr.funcName.text
@@ -1037,7 +1047,7 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileLit(expr: PartiqlAst.Expr.Lit, metas: MetaContainer): ThunkEnv {
+    internal open fun compileLit(expr: PartiqlAst.Expr.Lit, metas: MetaContainer): ThunkEnv {
         val value = ExprValue.of(expr.value.toIonValue(ion))
 
         return thunkFactory.thunkEnv(metas) { value }
@@ -1046,7 +1056,7 @@ internal class EvaluatingCompiler(
     private fun compileMissing(metas: MetaContainer): ThunkEnv =
         thunkFactory.thunkEnv(metas) { ExprValue.missingValue }
 
-    private fun compileId(expr: PartiqlAst.Expr.Id, metas: MetaContainer): ThunkEnv {
+    internal open fun compileId(expr: PartiqlAst.Expr.Id, metas: MetaContainer): ThunkEnv {
         val uniqueNameMeta = metas[UniqueNameMeta.TAG] as? UniqueNameMeta
         val fromSourceNames = currentCompilationContext.fromSourceNames
 
@@ -1177,7 +1187,7 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileIs(expr: PartiqlAst.Expr.IsType, metas: MetaContainer): ThunkEnv {
+    internal open fun compileIs(expr: PartiqlAst.Expr.IsType, metas: MetaContainer): ThunkEnv {
         val expThunk = compileAstExpr(expr.value)
         val typedOpParameter = expr.type.toTypedOpParameter()
         if (typedOpParameter.staticType is AnyType) {
@@ -1307,7 +1317,7 @@ internal class EvaluatingCompiler(
     private fun compileCast(expr: PartiqlAst.Expr.Cast, metas: MetaContainer): ThunkEnv =
         thunkFactory.thunkEnv(metas, compileCastHelper(expr.value, expr.asType, metas))
 
-    private fun compileCanCast(expr: PartiqlAst.Expr.CanCast, metas: MetaContainer): ThunkEnv {
+    internal open fun compileCanCast(expr: PartiqlAst.Expr.CanCast, metas: MetaContainer): ThunkEnv {
         val typedOpParameter = expr.asType.toTypedOpParameter()
         if (typedOpParameter.staticType is AnyType) {
             return thunkFactory.thunkEnv(metas) { ExprValue.newBoolean(true) }
@@ -1342,7 +1352,7 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileCanLosslessCast(expr: PartiqlAst.Expr.CanLosslessCast, metas: MetaContainer): ThunkEnv {
+    internal open fun compileCanLosslessCast(expr: PartiqlAst.Expr.CanLosslessCast, metas: MetaContainer): ThunkEnv {
         val typedOpParameter = expr.asType.toTypedOpParameter()
         if (typedOpParameter.staticType is AnyType) {
             return thunkFactory.thunkEnv(metas) { ExprValue.newBoolean(true) }
@@ -1412,7 +1422,7 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileSimpleCase(expr: PartiqlAst.Expr.SimpleCase, metas: MetaContainer): ThunkEnv {
+    internal open fun compileSimpleCase(expr: PartiqlAst.Expr.SimpleCase, metas: MetaContainer): ThunkEnv {
         val valueThunk = compileAstExpr(expr.expr)
         val branchThunks = expr.cases.pairs.map { Pair(compileAstExpr(it.first), compileAstExpr(it.second)) }
         val elseThunk = when (val default = expr.default) {
@@ -1445,7 +1455,7 @@ internal class EvaluatingCompiler(
         }
     }
 
-    private fun compileSearchedCase(expr: PartiqlAst.Expr.SearchedCase, metas: MetaContainer): ThunkEnv {
+    internal open fun compileSearchedCase(expr: PartiqlAst.Expr.SearchedCase, metas: MetaContainer): ThunkEnv {
         val branchThunks = expr.cases.pairs.map { compileAstExpr(it.first) to compileAstExpr(it.second) }
         val elseThunk = when (val default = expr.default) {
             null -> thunkFactory.thunkEnv(metas) { ExprValue.nullValue }
@@ -1743,7 +1753,15 @@ internal class EvaluatingCompiler(
         return offsetValue
     }
 
-    private fun compileSelect(selectExpr: PartiqlAst.Expr.Select, metas: MetaContainer): ThunkEnv {
+    internal open fun compileWhere(node: PartiqlAst.Expr): ThunkEnv {
+        return compileAstExpr(node)
+    }
+
+    internal open fun compileHaving(node: PartiqlAst.Expr): ThunkEnv {
+        return compileAstExpr(node)
+    }
+
+    internal open fun compileSelect(selectExpr: PartiqlAst.Expr.Select, metas: MetaContainer): ThunkEnv {
 
         // Get all the FROM source aliases and LET bindings for binding error checks
         val fold = object : PartiqlAst.VisitorFold<Set<String>>() {
@@ -1770,7 +1788,7 @@ internal class EvaluatingCompiler(
         return nestCompilationContext(ExpressionContext.NORMAL, emptySet()) {
             val fromSourceThunks = compileFromSources(selectExpr.from)
             val letSourceThunks = selectExpr.fromLet?.let { compileLetSources(it) }
-            val compiledWhere = selectExpr.where?.let { compileAstExpr(it) }
+            val compiledWhere = selectExpr.where?.let { compileWhere(it) }
             val sourceThunks = compileFromLetWhere(fromSourceThunks, letSourceThunks, compiledWhere)
 
             val orderByThunk = selectExpr.order?.let { compileOrderByExpression(it.sortSpecs) }
@@ -1924,7 +1942,7 @@ internal class EvaluatingCompiler(
                                     )
                                 }
 
-                                val havingThunk = selectExpr.having?.let { compileAstExpr(it) }
+                                val havingThunk = selectExpr.having?.let { compileHaving(it) }
 
                                 val filterHavingAndProject: (Environment, Group) -> ExprValue? =
                                     createFilterHavingAndProjectClosure(havingThunk, selectProjectionThunk)
@@ -3080,7 +3098,7 @@ private data class CompiledLetSource(
     val thunk: ThunkEnv
 )
 
-private enum class ExpressionContext {
+internal enum class ExpressionContext {
     /**
      * Indicates that the compiler is compiling a normal expression (i.e. not one of the other
      * contexts).